Event: process doc only if marker and active schema
